Introduce (semi-)automated process to source game-specific metadata from MAME
All games supported by MAME (up to 0.249) are supported by inp2json now,
and it should be easy to pull in support for games that become newly
supported by MAME in the future. The `--ports-count` argument also
becomes obsolete as a result.

Furthermore, a variety of other functional, stability- and
codestyle-related improvements is introduced.