Sanity-check that frame timestamps increase monotonically
If they do not, stop processing.

This catches the case where there is a discrepancy between the expected
count of ports and the actual count of ports in the INP payload. It does
not catch the case where the expected and actual counts of ports are the
same, but ports have been reordered at some point between the MAME
version the given INP file has been recorded with and the MAME version
`mame_inputport_ref.gz` has been generated against.

It is meant as a pragmatic fail-safe mechanism against incorrect output
in the former case.